CF PAT & README.md
