+++ cmdline/cmdline_suite_test.go
@@ -1,0 +1,13 @@
+package cmdline_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestPat(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Pat Suite")
+}
