--- output/csvs/1393026254904640835-98d93565-3c7c-4b6c-6be7-405204c622b0.csv
@@ -1,11 +1,0 @@
-Average,TotalTime,Total,TotalErrors,TotalWorkers,LastResult,WorstResult,WallTime,Type
-1000930820,1000930820,1,0,3,1000930820,1000930820,1001257797,0
-1500520898,3001041797,2,0,3,2000110977,2000110977,2000375974,0
-1667054609,5001163829,3,0,3,2000122032,2000122032,2000427735,0
-1500478858,6001915435,4,0,3,1000751606,2000122032,3001197049,0
-1800541166,9002705833,5,0,3,3000790398,3000790398,5001207272,0
-2167148223,13002889342,6,0,3,4000183509,4000183509,5001554402,0
-2286159355,16003115485,7,0,3,3000226143,4000183509,6001523865,0
-2250406246,18003249973,8,0,3,2000134488,4000183509,7001419150,0
-2444829469,22003465223,9,0,2,4000215250,4000215250,9001810868,0
-2500365524,25003655241,10,0,1,3000190018,4000215250,9001840917,0
