--- pat_suite_test.go
@@ -1,13 +1,0 @@
-package pat_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestPat(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Pat Suite")
-}
