+++ server/server_test.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"strings"
+	"os"
 
 	"github.com/julz/pat/config"
 	. "github.com/julz/pat/experiment"
@@ -29,6 +30,21 @@ var _ = Describe("Server", func() {
 		http.DefaultServeMux = http.NewServeMux()
 		ServeWithLab(lab)
 	})
+
+	AfterEach(func() {
+ 		os.Clearenv()
+ 	})
+
+ 	It("Checks if VCAP_APP_PORT exists and returns the port if true", func() {
+ 		os.Setenv("VCAP_APP_PORT", "1234")
+ 		port := GetPort()
+ 		Ω(port).Should(Equal("1234"))
+ 	})
+ 
+ 	It("Defaults to listening on port 8080 if the VCAP_APP_PORT environment variable is not set", func() {
+ 		port := GetPort()
+ 		Ω(port).Should(Equal("8080"))
+ 	})
 
 	It("Uses config to get CSV output directory", func() {
 		http.DefaultServeMux = http.NewServeMux()
